fix: project's responsiveness fixed
