+++ src/components/Wrapper.js
@@ -32,6 +32,10 @@ export const Wrapper = styled.section`
     padding: 60px 16px;
   }
 
+  @media screen and (max-width: 335px) {
+    padding: 60px 10px;
+  }
+
 
 
 `;
