+++ src/index.css
@@ -1,6 +1,5 @@
 * {
-  /* overflow-x: hidden;
-  width: 100%; */
+  width: 100%; 
 }
 
 body {
